+++ Page/Lists/ProductSearchList.php
@@ -10,6 +10,7 @@ use OxidEsales\Codeception\Module\Translation\Translator;
 use OxidEsales\Codeception\Page\Component\Header\AccountMenu;
 use OxidEsales\Codeception\Page\Component\Header\LanguageMenu;
 use OxidEsales\Codeception\Page\Component\Header\MiniBasket;
+use OxidEsales\Codeception\Page\Component\Header\SearchWidget;
 use OxidEsales\Codeception\Page\Details\ProductDetails;
 use OxidEsales\Codeception\Page\Page;
 
@@ -19,7 +20,7 @@ use OxidEsales\Codeception\Page\Page;
  */
 class ProductSearchList extends Page
 {
-    use LanguageMenu, MiniBasket, AccountMenu;
+    use LanguageMenu, MiniBasket, AccountMenu, SearchWidget;
 
     public $listItemTitle = '#searchList_%s';
 
@@ -32,6 +33,18 @@ class ProductSearchList extends Page
     public $variantSelection = '#variantselector_searchList_%s button';
 
     public $sortingSelection = '//a[@title="%s"]';
+
+    public $itemsPerPageSelection = '//div[@class="btn-group open"]//*[contains(text(),"%s")]';
+
+    public $listViewSelection = '//ul[@class="dropdown-menu"]//*[contains(text(),"%s")]';
+
+    public $nextListPage = '//ol[@id="itemsPager"]/li[@class="next"]/a';
+
+    public $previousListPage = '//ol[@id="itemsPager"]/li[@class="prev"]/a';
+
+    public $pageNumberSelection = '//ol[@id="itemsPager"]//a[contains(text(),"%s")]';
+
+    public $activePageNumber = '//ol[@id="itemsPager"]/li[@class="active"]/a[contains(text(),"%s")]';
 
     /**
      * @param mixed $param
@@ -58,6 +71,18 @@ class ProductSearchList extends Page
         $I->see($productData['title'], sprintf($this->listItemTitle, $itemId));
         $I->see($productData['description'], sprintf($this->listItemDescription, $itemId));
         $I->see($productData['price'], sprintf($this->listItemPrice, $itemId));
+        return $this;
+    }
+
+    /**
+     * @param int $count
+     *
+     * @return $this
+     */
+    public function seeSearchCount(int $count)
+    {
+        $I = $this->user;
+        $I->see($count . ' ' . Translator::translate('HITS_FOR'));
         return $this;
     }
 
@@ -112,7 +137,76 @@ class ProductSearchList extends Page
     {
         $I = $this->user;
         $I->click(Translator::translate('SORT_BY'));
+        $I->waitForElement(sprintf($this->sortingSelection, $this->getSortingElementTitle($sortingName, $sortingOrder)));
         $I->click(sprintf($this->sortingSelection, $this->getSortingElementTitle($sortingName, $sortingOrder)));
+
+        return $this;
+    }
+
+    /**
+     * @param int $item
+     *
+     * @return ProductSearchList
+     */
+    public function selectProductsPerPage(int $item): ProductSearchList
+    {
+        $I = $this->user;
+        $I->click(Translator::translate('PRODUCTS_PER_PAGE'));
+        $I->click(sprintf($this->itemsPerPageSelection, $item));
+        $I->waitForText(Translator::translate('PRODUCTS_PER_PAGE') . ' ' . $item);
+
+        return $this;
+    }
+
+    /**
+     * @param string $view
+     *
+     * @return ProductSearchList
+     */
+    public function selectListDisplayType(string $view): ProductSearchList
+    {
+        $I = $this->user;
+        $I->click(Translator::translate('LIST_DISPLAY_TYPE'));
+        $I->click(sprintf($this->listViewSelection, $view));
+        $I->waitForText(Translator::translate('LIST_DISPLAY_TYPE') . ' ' . $view);
+
+        return $this;
+    }
+
+    /**
+     * @return ProductSearchList
+     */
+    public function openNextListPage(): ProductSearchList
+    {
+        $I = $this->user;
+        $I->click($this->nextListPage);
+        $I->waitForPageLoad();
+
+        return $this;
+    }
+
+    /**
+     * @return ProductSearchList
+     */
+    public function openPreviousListPage(): ProductSearchList
+    {
+        $I = $this->user;
+        $I->click($this->previousListPage);
+        $I->waitForPageLoad();
+
+        return $this;
+    }
+
+    /**
+     * @param int $pageNumber
+     *
+     * @return ProductSearchList
+     */
+    public function openListPageNumber(int $pageNumber): ProductSearchList
+    {
+        $I = $this->user;
+        $I->click(sprintf($this->pageNumberSelection, $pageNumber));
+        $I->waitForElement(sprintf($this->activePageNumber, $pageNumber));
 
         return $this;
     }